backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
supabase==1.0.3
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
openai==0.27.8
python-dotenv==1.0.0
lxml[html_clean]==4.9.3
numpy
# TODO: Add more dependencies as needed (e.g., for scraping, embeddings, etc.)